backend/requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
python-dotenv
pydantic>=2.0
pydantic-settings
